feat: hub crud 개발

+++ hub/src/main/java/com/qpang/hub/presentation/dto/HubResponse.java
@@ -2,10 +2,16 @@ package com.qpang.hub.presentation.dto;
 
 import com.qpang.hub.domain.model.Hub;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record HubResponse(
-        UUID id, String name, String address, Double latitude, Double longitude, UUID managerId
+        UUID id,
+        String name,
+        String address,
+        BigDecimal latitude,
+        BigDecimal longitude,
+        UUID managerId
 ) {
     public static HubResponse from(Hub hub) {
         return new HubResponse(hub.getId(), hub.getName(), hub.getAddress(),
